// test/helper.test.d.ts
export enum FaultType {
    SenderAsReserver,
    DispatcherAsReserver,
    SenderAsDispatcher,
    DispatcherAsReceiver,
    ReceiverAsReserver,
    ReserverAsSender,
    ReserverAsReceiver
}

export enum DeliveryStepType {
    DeliveryByDispatcherFromReserverToReceiver,
    DeliveryByDispatcherFromReserverToReserver,
    DeliveryByDispatcherFromSenderToReceiver,
    DeliveryByDispatcherFromSenderToReserver,
    SelfDeliveryBySenderToReserver
}